Move salesforce authenticator into a lib, and NamedError into the lib folder
