+++ src/lib/NamedError.js
